type change

+++ basis.h
@@ -36,7 +36,7 @@ class   CBasisFuncPos
 {
 private:
 	bool is_try_find;
-	int pos_saved;
+	ptrdiff_t pos_saved;
 
 public:
 	CBasisFuncPos() : is_try_find(false) , pos_saved(-1) { }
@@ -46,7 +46,7 @@ public:
 
 	bool IsNotFound() { return is_try_find && (pos_saved == -1); }
 
-	int Find(CAyaVM &vm,const yaya::char_t *name);
+	ptrdiff_t Find(CAyaVM &vm,const yaya::char_t *name);
 };
 
 class	CBasis
